Deprecating :selector_method parameter
The Nokogiri API allows for the client to use `:search`, which will perform either a CSS or XPath node traversal depending on the syntax of the string. For now, I've made the references to :selector_method innocuous...whether the user specifies :css or :xpath, only :search is invoked. On the next iteration, :selector_method should be removed as the third argument.